Refactor common function
